Move DI stuff into folder
