+++ src/ForEvolve.ExceptionFilters.Abstraction/DependencyInjection/IExceptionMappingBuilder.cs
@@ -1,0 +1,7 @@
+﻿namespace Microsoft.Extensions.DependencyInjection
+{
+    public interface IExceptionMappingBuilder
+    {
+        IServiceCollection Services { get; }
+    }
+}
