+++ src/ForEvolve.ExceptionFilters.Abstraction/DependencyInjection/ServiceCollectionExtensions.cs
@@ -1,0 +1,18 @@
+﻿using ForEvolve.ExceptionFilters;
+using Microsoft.Extensions.DependencyInjection.Extensions;
+using Microsoft.Extensions.Options;
+using System;
+
+namespace Microsoft.Extensions.DependencyInjection
+{
+    public static class ServiceCollectionExceptionFiltersExtensions
+    {
+        public static IServiceCollection AddExceptionMapping(this IServiceCollection services, Action<IExceptionMappingBuilder> exceptionMappingBuilder = null)
+        {
+            services.AddLogging();
+            exceptionMappingBuilder?.Invoke(new ServiceCollectionWrapper(services));
+            services.TryAddSingleton<IExceptionHandlingManager, ExceptionHandlingManager>();
+            return services;
+        }
+    }
+}
