+++ src/ForEvolve.ExceptionFilters.Abstraction/DependencyInjection/ServiceCollectionWrapper.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace Microsoft.Extensions.DependencyInjection
+{
+    public class ServiceCollectionWrapper : IExceptionMappingBuilder
+    {
+        public ServiceCollectionWrapper(IServiceCollection services)
+        {
+            Services = services ?? throw new ArgumentNullException(nameof(services));
+        }
+
+        public IServiceCollection Services { get; }
+    }
+}
